[Vita] Implement proper __malloc_lock, __malloc_unlock
