+++ newlib/libc/sys/vita/Makefile.am
@@ -8,7 +8,7 @@ AM_CCASFLAGS = $(INCLUDES)
 
 noinst_LIBRARIES = lib.a
 
-lib_a_SOURCES = syscalls.c sbrk.c threading.c reent.c
+lib_a_SOURCES = syscalls.c sbrk.c threading.c mlock.c
 lib_a_CCASFLAGS = $(AM_CCASFLAGS)
 lib_a_CFLAGS = $(AM_CFLAGS)
